+++ src/redsun_mimir/__main__.py
@@ -17,24 +17,12 @@ def main() -> None:
     subparsers.add_parser(
         "uc2", help="Run UC2 microscope application with pre-shipped configuration."
     )
-    subparsers.add_parser("motor", help="Run the mock motor example")
-    subparsers.add_parser("light", help="Run the mock light example")
-    subparsers.add_parser(
-        "acquisition",
-        help="Run the example acquisition container",
-    )
 
     options = parser.parse_args(namespace=Options())
     if options.command == "sim":
         configurations.run_simulation_container()
     elif options.command == "uc2":
         configurations.run_uc2_container()
-    elif options.command == "motor":
-        configurations.run_stage_container()
-    elif options.command == "light":
-        configurations.run_light_container()
-    elif options.command == "acquisition":
-        configurations.run_acquisition_container()
     else:
         parser.print_help()